Split up pages
Created a new file for each of the application's pages.
